Added refactoring TODO document and updated constants for time periods and chart thresholds
- Created REFACTORING_TODO.md to outline refactoring tasks for the Stats app.
- Introduced TIME_PERIODS and CHART_THRESHOLDS constants in constants.ts for better maintainability.
- Updated useGrowthStats and chart-helpers to utilize new constants for improved readability and consistency.
- Refactored GrowthKPIs component to leverage useGrowthChartData hook for chart data management.
- Simplified Overview component by integrating useOverviewChartData hook for data processing.
