+++ apps/stats/src/views/Stats/Growth/components/GrowthKPIs.tsx
@@ -1,10 +1,11 @@
-import React, {useEffect, useMemo, useState} from 'react';
-import {BarChartLoadingIndicator, ChartConfig, ChartContainer, ChartTooltip, ChartTooltipContent, GhAreaChart, GhAreaChartDataItem, KpiTabTrigger, KpiTabValue, Recharts, Tabs, TabsContent, TabsList, TabsTrigger, centsToDollars, formatDisplayDateWithRange, formatNumber} from '@tryghost/shade';
+import React, {useEffect, useState} from 'react';
+import {BarChartLoadingIndicator, ChartContainer, ChartTooltip, ChartTooltipContent, GhAreaChart, KpiTabTrigger, KpiTabValue, Recharts, Tabs, TabsContent, TabsList, TabsTrigger, centsToDollars, formatNumber} from '@tryghost/shade';
 import {DiffDirection} from '@src/hooks/useGrowthStats';
+import {GROWTH_TAB_CONFIG, PAID_CHANGE_CHART_CONFIG} from '../constants';
 import {STATS_RANGES} from '@src/utils/constants';
-import {sanitizeChartData} from '@src/utils/chart-helpers';
 import {useAppContext} from '@src/App';
 import {useGlobalData} from '@src/providers/GlobalDataProvider';
+import {useGrowthChartData} from '../hooks/useGrowthChartData';
 import {useNavigate, useSearchParams} from '@tryghost/admin-x-framework';
 
 type ChartDataItem = {
@@ -67,146 +68,7 @@ const GrowthKPIs: React.FC<{
     };
 
     const {totalMembers, freeMembers, paidMembers, mrr, percentChanges, directions} = totals;
-
-    // Create chart data based on selected tab
-    const chartData = useMemo(() => {
-        if (!allChartData || allChartData.length === 0) {
-            return [];
-        }
-
-        // First sanitize the data based on the selected field
-        let sanitizedData: ChartDataItem[] = [];
-        let fieldName: keyof ChartDataItem = 'value';
-
-        switch (currentTab) {
-        case 'free-members':
-            fieldName = 'free';
-            break;
-        case 'paid-members':
-            fieldName = 'paid';
-            break;
-        case 'mrr': {
-            fieldName = 'mrr';
-            break;
-        }
-        default:
-            fieldName = 'value';
-        }
-
-        sanitizedData = sanitizeChartData(allChartData, range, fieldName, 'exact');
-
-        // Then map the sanitized data to the final format
-        let processedData: GhAreaChartDataItem[] = [];
-
-        switch (currentTab) {
-        case 'free-members':
-            processedData = sanitizedData.map((item, index) => {
-                const diffValue = index === 0 ? null : item.free - sanitizedData[index - 1].free;
-                return {
-                    ...item,
-                    value: item.free,
-                    formattedValue: formatNumber(item.free),
-                    label: 'Free members',
-                    diffValue,
-                    formattedDiffValue: diffValue === null ? null : (diffValue < 0 ? `-${formatNumber(diffValue)}` : `+${formatNumber(diffValue)}`)
-                };
-            });
-            break;
-        case 'paid-members':
-            processedData = sanitizedData.map((item, index) => {
-                const diffValue = index === 0 ? null : item.paid - sanitizedData[index - 1].paid;
-                return {
-                    ...item,
-                    value: item.paid,
-                    formattedValue: formatNumber(item.paid),
-                    label: 'Paid members',
-                    diffValue,
-                    formattedDiffValue: diffValue === null ? null : (diffValue < 0 ? `-${formatNumber(diffValue)}` : `+${formatNumber(diffValue)}`)
-                };
-            });
-            break;
-        case 'mrr':
-            processedData = sanitizedData.map((item, index) => {
-                const diffValue = index === 0 ? null : centsToDollars(item.mrr) - centsToDollars(sanitizedData[index - 1].mrr);
-                return {
-                    ...item,
-                    value: centsToDollars(item.mrr),
-                    formattedValue: `${currencySymbol}${formatNumber(centsToDollars(item.mrr))}`,
-                    label: 'MRR',
-                    diffValue,
-                    formattedDiffValue: diffValue === null ? null : (diffValue < 0 ? `-${currencySymbol}${formatNumber(diffValue * -1)}` : `+${currencySymbol}${formatNumber(diffValue)}`)
-                };
-            });
-            break;
-        default:
-            processedData = sanitizedData.map((item, index) => {
-                const currentTotal = item.free + item.paid + item.comped;
-                const previousTotal = index === 0 ? null : sanitizedData[index - 1].free + sanitizedData[index - 1].paid + sanitizedData[index - 1].comped;
-                const diffValue = index === 0 ? null : currentTotal - previousTotal!;
-                return {
-                    ...item,
-                    value: currentTotal,
-                    formattedValue: formatNumber(currentTotal),
-                    label: 'Total members',
-                    diffValue,
-                    formattedDiffValue: diffValue === null ? null : (diffValue < 0 ? `-${formatNumber(diffValue)}` : `+${formatNumber(diffValue)}`)
-                };
-            });
-        }
-
-        return processedData;
-    }, [currentTab, allChartData, range, currencySymbol]);
-
-    const tabConfig = {
-        'total-members': {
-            color: 'hsl(var(--chart-darkblue))'
-        },
-        'free-members': {
-            color: 'hsl(var(--chart-blue))'
-        },
-        'paid-members': {
-            color: 'hsl(var(--chart-purple))'
-        },
-        mrr: {
-            color: 'hsl(var(--chart-teal))'
-        }
-    };
-
-    const paidChangeChartData = useMemo(() => {
-        if (currentTab !== 'paid-members') {
-            return [];
-        }
-
-        if (!allChartData || allChartData.length === 0) {
-            return [];
-        }
-
-        // First sanitize the data for the current range
-        const sanitizedData = sanitizeChartData(allChartData, range, 'paid', 'exact');
-
-        // Transform the sanitized data into the format expected by the chart
-        return sanitizedData.map((item) => {
-            // Format date in a more readable format (e.g., "25 May")
-            const date = new Date(item.date);
-
-            return {
-                date: formatDisplayDateWithRange(date, range),
-                new: item.paid_subscribed || 0,
-                cancelled: -(item.paid_canceled || 0) // Negative for the stacked bar chart
-            };
-        });
-    }, [currentTab, allChartData, range]);
-
-    const paidChangeChartConfig = {
-        new: {
-            label: 'New',
-            color: 'hsl(var(--chart-teal))'
-        },
-        cancelled: {
-            label: 'Cancelled',
-            color: 'hsl(var(--chart-rose))'
-        }
-    } satisfies ChartConfig;
+    const {chartData, paidChangeChartData} = useGrowthChartData(currentTab, allChartData, range, currencySymbol);
 
     if (isLoading) {
         return (
@@ -295,7 +157,7 @@ const GrowthKPIs: React.FC<{
                         <TabsContent value="total">
                             <GhAreaChart
                                 className={areaChartClassname}
-                                color={tabConfig[currentTab as keyof typeof tabConfig].color}
+                                color={GROWTH_TAB_CONFIG[currentTab as keyof typeof GROWTH_TAB_CONFIG].color}
                                 data={chartData}
                                 dataFormatter={formatNumber}
                                 id="paid-members"
@@ -303,7 +165,7 @@ const GrowthKPIs: React.FC<{
                             />
                         </TabsContent>
                         <TabsContent value="change">
-                            <ChartContainer className='mt-6 max-h-[280px] w-full' config={paidChangeChartConfig}>
+                            <ChartContainer className='mt-6 max-h-[280px] w-full' config={PAID_CHANGE_CHART_CONFIG}>
                                 <Recharts.BarChart
                                     data={paidChangeChartData}
                                     stackOffset='sign'
@@ -363,7 +225,7 @@ const GrowthKPIs: React.FC<{
                                                                     } as React.CSSProperties}
                                                                 />
                                                                 <span className='text-sm text-muted-foreground'>
-                                                                    {paidChangeChartConfig[name as keyof typeof paidChangeChartConfig]?.label || name}
+                                                                    {PAID_CHANGE_CHART_CONFIG[name as keyof typeof PAID_CHANGE_CHART_CONFIG]?.label || name}
                                                                 </span>
                                                             </div>
                                                             <div className="ml-auto flex items-baseline gap-0.5 font-mono font-medium tabular-nums text-foreground">
@@ -404,7 +266,7 @@ const GrowthKPIs: React.FC<{
                                 <div className='flex items-center gap-1'>
                                     <span className='size-2 rounded-full opacity-50'
                                         style={{
-                                            backgroundColor: paidChangeChartConfig.new.color
+                                            backgroundColor: PAID_CHANGE_CHART_CONFIG.new.color
                                         }}
                                     ></span>
                                     New
@@ -412,7 +274,7 @@ const GrowthKPIs: React.FC<{
                                 <div className='flex items-center gap-1'>
                                     <span className='size-2 rounded-full opacity-50'
                                         style={{
-                                            backgroundColor: paidChangeChartConfig.cancelled.color
+                                            backgroundColor: PAID_CHANGE_CHART_CONFIG.cancelled.color
                                         }}
                                     ></span>
                                     Cancelled
